reorganize arxiv logic for using as management commands
